Migrated the SignUp functionality into the SignUpForm Component.
